# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API - exemplo",
    "description": "API de exemplo com dados de cursos, professores e disciplinas"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Login",
        "description": "Este endpoint faz o login",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "senha": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/cursos/": {
      "get": {
        "tags": [
          "Cursos"
        ],
        "summary": "Retorna a lista dos cursos",
        "description": "Este endpoint retorna uma lista dos cursos",
        "responses": {
          "200": {
            "description": "Lista dos cursos",
            "content": {
              "application/json": {
                "schema": {
                  "array": {
                    "$ref": "#/components/schemas/curso"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cursos"
        ],
        "summary": "Insere um curso",
        "description": "Este endpoint insere um novo curso",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/curso"
              }
            }
          }
        }
      }
    },
    "/cursos/{id}": {
      "get": {
        "tags": [
          "Cursos"
        ],
        "summary": "Retorna um curso pelo id",
        "description": "Este endpoint retorna um curso, dado um id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do curso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/curso"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cursos"
        ],
        "summary": "Altera os dados de um curso",
        "description": "Este endpoint altera os dados de um curso",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Cursos"
        ],
        "summary": "Exclui um curso",
        "description": "Este endpoint excluir um curso",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/disciplinas/": {
      "get": {
        "tags": [
          "Disciplinas"
        ],
        "summary": "Retorna a lista das disciplinas",
        "description": "Este endpoint retorna uma lista das discipinas",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Disciplinas"
        ],
        "summary": "Insere disciplina",
        "description": "Este endpoint insere uma nova disciplina",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/disciplinas/{id}": {
      "get": {
        "tags": [
          "Disciplinas"
        ],
        "summary": "Retorna os dados de uma disciplina",
        "description": "Este endpoint retorna os dados de uma disciplina, dado o seu id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Disciplinas"
        ],
        "summary": "Altera os dados disciplina",
        "description": "Este endpoint altera os dados de uma disciplina",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Disciplinas"
        ],
        "summary": "Exclui disciplina",
        "description": "Este endpoint exclui uma disciplina",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/professores/": {
      "get": {
        "tags": [
          "Professores"
        ],
        "summary": "Retorna a lista dos professores",
        "description": "Este endpoint retorna uma lista dos professores",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Professores"
        ],
        "summary": "Insere professor",
        "description": "Este endpoint insere um professor",
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/professores/{id}": {
      "get": {
        "tags": [
          "Professores"
        ],
        "summary": "Retorna um professor pelo id",
        "description": "Este endpoint retorna os dados de um professor, dado o seu id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Professores"
        ],
        "summary": "Altera os dados de um professor",
        "description": "Este endpoint altera os dados de um professor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Professores"
        ],
        "summary": "Exclui um professor",
        "description": "Este endpoint exclui um professor",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "professor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Rafael"
          },
          "email": {
            "type": "string",
            "example": "rafael.speroni@...br"
          }
        }
      },
      "curso": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "nome": {
            "type": "string",
            "example": "Tecnologia em Sistemas para Internet"
          },
          "sigla": {
            "type": "string",
            "example": "TSI"
          },
          "descricao": {
            "type": "string",
            "example": "..."
          },
          "coordenador": {
            "type": "number",
            "example": 1
          }
        }
      }
    }
  }
}